RDKCOM-2906: REFPLTV-1146 Bring up 64bit support in Rpi4 (for both kernel an...
Reason for change: to match datatype compatibilty
Test Procedure: Build should be successful. ctrl-main should build fine.
Risks: Low/None.

(cherry picked from commit a7c5935d6fee279413452349e52ed76416352ac3)

+++ src/ctrlm_voice_session.cpp
@@ -171,7 +171,7 @@ static int vrex_response_timeout(void *data) {
 
 int vrex_request_progress(void *clientp, curl_off_t dltotal, curl_off_t dlnow, curl_off_t ultotal, curl_off_t ulnow) {
    voice_session_t *session = (voice_session_t *)clientp;
-   if(session->abort_vrex_request_get() && session->voice_data_qty_sent_get() <= ulnow) {
+   if(session->abort_vrex_request_get() && session->voice_data_qty_sent_get() <= (long unsigned int)ulnow) {
       LOG_INFO("%s: Aborting CURL session due to vrex response timeout\n", __FUNCTION__);
       session->abort_vrex_request_set(false);
       session->vrex_response_timeout_reset();
